refactor: move requestDTOs to new package
